filejournal: add a min time interval between writes
If the journal devices has NVRAM, it may end up giving us near-zero latency
for small writes, which means we will write each event individually.  For
a stream of small writes, this might be suboptimal.

This patch will force us to wait some minimum amount of time between
dispatching a new write to the journal device.

Leave this off by default, but add the tunable.

Signed-off-by: Sage Weil <sage@inktank.com>