mon: Monitor: enforce features on converted store
Check if the store being converted has been converted to v0.52, and if it
has the GV feature set. We require the global versions on all paxos
machines before we can start the conversion, so we must make sure this
feature is set on disk.

Also, upon finishing the conversion, we must then remove the GV feature
and set our Single-Paxos feature; otherwise, our monitor will not start.

Signed-off-by: Joao Eduardo Luis <joao.luis@inktank.com>