filestore: introduce fd cache
Introduce a cache of recent file descriptors.  This avoids the overhead of
doing a path traversal and kernel open(2) call when, in many cases, we are
doing a series of operations on the same file.

This has a noticeable performance impact for fast storages (SSDs, NVRAM,
ramdisk) and small write workloads (e.g., small, sequential writes).

Signed-off-by: Sage Weil <sage@inktank.com>