crush: adjust default tunables
Make the default for new crush maps use improved values for the
tunables.  This makes newly created maps incompatible with older code
that does not support them, which means (for example) that some
pre-bobtail code and older kernels may not be able to communicate with
newly created clusters.  Anything bobtail or later, however, will be
fine.

Signed-off-by: Sage Weil <sage@inktank.com>