rgw: don't schedule a cr if already scheduled
This can be triggered when multiple IOs return for the same cr. We
don't want to allow it to be scheduled multiple times, as this
might wake up blocking IOs that shouldn't be woken up. If we try to
block on an IO that we already complete, we won't block anyway.

Signed-off-by: Yehuda Sadeh <yehuda@redhat.com>