rgw: pubsub: initial topic datastructure implementation
Need to implement a scalable repository for topic events. The
events will be stored in a per-topic cls_log controlled omap
that will guard the max number of entries. Once the space in the
omap is exhausted we'll start populating a new one. Will have
multiple shards per topic (each shard could have multiple logs
as described above). We'll store the id of the last log obj of
each shard in a special meta object.

Signed-off-by: Yehuda Sadeh <yehuda@redhat.com>