rgw: bucket sync: split sip cr managers
Similar to what was done with data sync, creating two separte sip cr
managers instead of a single one that has different calls for full
and inc. This required making the marker tracker for full sync use
a string instead of rgw_obj_key, and also creating a common
representation of the fetched entries. Will make the transition
to sip handlers easier.

Signed-off-by: Yehuda Sadeh <yehuda@redhat.com>