rgw: adds a preemptive mode to rgw's dynamic resharding
Our buckets in production tend to follow an exponential distribution with
respect to object count. Since we have relatively few large buckets, and
we can only use dynamic bucket resharding on buckets with a couple
hundred thousand objects without causing a significant request backlog,
this patch makes it possible to configure RGW to dynamically reshard the
largest of the small buckets within a given size range to a statically
configured number of shards. This allows small buckets to grow to an
arbitrary size without needing to pause reads, potentially many times,
using RGW's normal dynamic resharding process.

Note that buckets will be queued for dynamic resharding as objects are
added to a given bucket and that bucket's usage info is checked against
its current number of shards. Buckets should never "escape" dynamic
resharding, growing too large for the configured resharding range. It is
possible a bucket will be resharded once it has more objects than the
given range, if the range is especially small or the reshard queue (aka
log) develops a backlog of work.

In addition to the options for dynamic resharding, this patch reduces
the minimum value for rgw_reshard_thread_interval. This interval
determines how long RGW will pause before looking at the reshard queue,
and its 10 minutes is probably too long for preemptive resharding.

Signed-off-by: Nick Janus <njanus@digitalocean.com>