ceph-volume: Don't check for locked raw devices.
This was added as a heuristic when ceph-volume enumerates possible OSD
candidates. It's not useful in our environment, since any such sharing
would be explicit, and actually gets in the way when 'simple scan' is
run in parallel, since the os.close() here causes a partition rescan
which can disrupt other running instances.