rgw: rgw_obj_key encodes snap id in instance field
snap_id is needed when bucket snapshots are enabled, but not
when instance exists (as in versioning enabled), as the versioned
objects are not going to have different instances. In the null version
case we need to be able to differentiate between the different instances
(as in different snapshots). The rados object name overrides the
instance string in that case.

Signed-off-by: Yehuda Sadeh <ysadehwe@ibm.com>