rgw: modify the target obj to hold the snapshot id
When we're dealing with null version object that has a snapshot-id, we
need to allow it to be overwritten. This is different than how regular
versioned objects work, because for these we generate new version id
for every object instance. The problem is that since the target->obj
doesn't have the snapshot id, the atomic checks are based on the olh
info and not on the object instance info. Letting the target->obj being
updated allows the atomic tests to be done on the correct entity.

Signed-off-by: Yehuda Sadeh <ysadehwe@ibm.com>