rgw: rework object snap representation
Beforehand the snap_id field in the rgw_obj_key to identify whether an
object instance belong to a specific snapshot. However, this was inconsistent
with the idea that we don't want to physically represent versioned objects
(as in non-nulled) without the snap_id. This lead to a lot of inconsistencies.
Now rgw_obj_key only represents the object instance key: versioned objects
will not have snap_id set, even if they're in a snapshot. Non-versioned
objects will have the snap_id set. The RGWObjState now has a snap_id
field, and a snap_id attr was added to the object's meta.

Signed-off-by: Yehuda Sadeh <ysadehwe@ibm.com>