cls/rgw: snap_id key related fixes
The key may or may not have snap_id set. When reading existing object, use
the key to get to the correct index entries. When creating new one, we
need to use the snap_id from the op.
A bucket dir entry now encodes snap_id as part of its meta, and also encodes
the key's snap_id separately (as these might not hold the same value -- the
key, if pointing to a versioned object, might not have that field set).

Signed-off-by: Yehuda Sadeh <ysadehwe@ibm.com>