rgw: delete object snapshot snapshot related changes
The delete object op has a few cases:
if version_id is not specified it is a delete marker creation operation
(can be either versioned or non-versioned).
If version id is specified then it's an object instance removal. In which
case there are two cases (null versioned, versioned).
In the null versioned removal case we need to pass a key with instance set
to 'null', but also with a snap_id.
In a null-versioned delete marker creation we need to have a key with
an empty instance and no snap_id.
When it's non-null object, the snap_id field should not be set (and it's
assumed that it's not set).

Signed-off-by: Yehuda Sadeh <ysadehwe@ibm.com>