cls/rgw: track current null version snap_id in olh
In regular operation (when snapshots aren't enabled), the null
versions are implicitly removed. However, when adding snapshots
if the null versioned object has a different snapshot id than
the previous one, we keep separate new entries for it. However,
we want to mark the null versioned object that belongs to the
previous snapshot as "removed at" the current snapshot, so that
we know when this object has been "overwritten". This is used
in order for us to know when an object can be actually removed,
or whether it's still alive.

Signed-off-by: Yehuda Sadeh <ysadehwe@ibm.com>