cls/rgw: support for per-snapshot bucket stats
Adding support for per-snapshot bucket stats. We now kee a new
per-snapshot index that keeps both the aggregated stats (for that
snapshot), and the specific snapshot stats.
These stats are only written after a newer snapshot is written, so that
we don't need to write to another index key on every write. We keep
extra fields in the bucket header to store the current snapshot stats
while its being written.

Signed-off-by: Yehuda Sadeh <ysadehwe@ibm.com>